requirements.txt
fastapi
requests
uvicorn
python-jose
bcrypt==4.0.1
passlib